refactor: types.d.ts를 types폴더로 이동
